+++ lizard_map/views.py
@@ -5,8 +5,11 @@ except ImportError:
 import csv
 import datetime
 import logging
+import math
 import re
 import urllib2
+from xml.dom.minidom import parseString
+
 from dateutil import parser as date_parser
 
 from PIL import Image
@@ -1171,7 +1174,7 @@ def wms(request, workspace_item_id, workspace_storage_id=None,
     return response
 
 
-def search(workspace, google_x, google_y, radius):
+def search(workspace, google_x, google_y, radius, request=None):
     """Search workspace for given coordinates.
 
     Return a list of found results in "adapter.search" dictionary
@@ -1182,9 +1185,13 @@ def search(workspace, google_x, google_y, radius):
     for workspace_item in workspace.workspace_items.filter(
         visible=True):
 
+        requestarg = {}
+        if getattr(workspace_item.adapter, 'search_with_request', False):
+            requestarg = {'request': request}
+
         try:
             search_results = workspace_item.adapter.search(
-                google_x, google_y, radius=radius)
+                google_x, google_y, radius=radius, **requestarg)
             found += search_results
         except:
             logger.exception(
@@ -1217,8 +1224,10 @@ def search_coordinates(request,
     y = float(request.GET.get('y'))
     format = request.GET.get('format', _format)
 
+    top = float(request.GET.get('extent_top'))
+    bottom = float(request.GET.get('extent_bottom'))
     # TODO: convert radius to correct scale (works now for google + rd)
-    radius = float(request.GET.get('radius'))
+    radius = math.fabs(top - bottom) / 30
     radius_search = radius
 
     if 'HTTP_USER_AGENT' in request.META:
@@ -1247,7 +1256,7 @@ def search_coordinates(request,
 
     # The first actual search
     # popup_json() does the second real search
-    found = search(workspace, google_x, google_y, radius)
+    found = search(workspace, google_x, google_y, radius, request=request)
     logger.debug('>>> FOUND <<< %s\n%s' % (format, repr(found)))
 
     if found:
@@ -1256,21 +1265,22 @@ def search_coordinates(request,
         if format == 'name':
             result = {}
             result['name'] = found[0]['name']
+
             # For the x/y we use the original x/y value to position
             # the popup to the lower right of the cursor to prevent
             # click propagation problems.
             result['x'] = x + (radius / 10)
             result['y'] = y - (radius / 10)
             return HttpResponse(json.dumps(result))
-        elif format == 'object':
-            result = [{'id':f['identifier'], 'name':f['name']}
+        if format == 'object':
+            result = [{'id': f['identifier'], 'name': f['name']}
                       for f in found]
-
             return HttpResponse(json.dumps(result))
-         # default: as popup
+
+        # default: as popup
         return popup_json(found, request=request)
     if format == 'object':
-        return HttpResponse([])
+            return HttpResponse([])
     return popup_json([], request=request)
 
 
@@ -1367,7 +1377,7 @@ class CollageView(CollageMixin, ActionDialogView):
 
         collage = get_collage_edit_by_request(self.request)
 
-        found = search(workspace, google_x, google_y, radius)
+        found = search(workspace, google_x, google_y, radius, self.request)
 
         for found_item in found:
             # Add all found items to collage.